Edit X,Y offset selection to be  blacklist-based
I found with further testing that thunderbird, pidgin, paprefs, and
pactl also were being shifted by X,Y amount equal to window decorations.
Knowing this, it seems that konsole is my only exception for some
reason, so I now shift all windows with the exception of processes pith
name in a blacklist. I use `ps -p` to find the process name for an
associated PID and save that info.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>